fix: Delete old plan file in GCS when terraform plan fails

+++ terraform-plan/main.sh
@@ -11,7 +11,12 @@ set -e
 
 if [ "$code" = "1" ]; then
 	echo "===> Delete old plan file to prevent the accident" >&2
-	aws s3 rm "s3://$S3_BUCKET_NAME_PLAN_FILE/$PR_NUMBER/$TFACTION_TARGET/tfplan.binary" || :
+	if [ -n "${S3_BUCKET_NAME_PLAN_FILE:-}" ]; then
+		aws s3 rm "s3://$S3_BUCKET_NAME_PLAN_FILE/$PR_NUMBER/$TFACTION_TARGET/tfplan.binary" || :
+	fi
+	if [ -n "${GCS_BUCKET_NAME_PLAN_FILE:-}" ]; then
+		gsutil rm "gs://$GCS_BUCKET_NAME_PLAN_FILE/$PR_NUMBER/$TFACTION_TARGET/tfplan.binary" || :
+	fi
 	exit 1
 fi
